pyproject.toml
[tool.black]
line-length = 100
target-version = ['py39']
include = '\.py$'
exclude = '''
/(
  | \.git
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.poetry]
name = "thesis_nlp"
version = "0.0.0"
description = ""
authors = ["Antoni Maciag <antoni.maciag1@gmail.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "torchcpu"
url = "https://download.pytorch.org/whl/cpu/"
priority = "supplemental"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = "~3.9"

black = "*"
flake8 = "*"
isort = "*"
mypy = "*"
pandas = "*"
pyyaml = "*"
torch = "2.1.1"
numpy= "1.22.2"
huggingface-hub = "0.23.0"
transformers = "4.38"
datasets = "2.18"
sentence_transformers = "*"
matplotlib = "*"
nltk = "*"
protobuf = "*"
textattack = "0.3.9"
selenium = "*"